exposed metrics and provider on spaceAbout, not on Space
