+++ src/domain/space/space.about/space.about.resolver.fields.ts
@@ -10,10 +10,17 @@ import { Loader } from '@core/dataloader/decorators/data.loader.decorator';
 import { ILoader } from '@core/dataloader/loader.interface';
 import { ISpaceAbout } from './space.about.interface';
 import { SpaceAbout } from './space.about.entity';
+import { INVP } from '@domain/common/nvp/nvp.interface';
+import { SpaceAboutService } from './space.about.service';
+import { SpaceLookupService } from '../space.lookup/space.lookup.service';
+import { IContributor } from '@domain/community/contributor/contributor.interface';
 
 @Resolver(() => ISpaceAbout)
 export class SpaceAboutResolverFields {
-  constructor() {}
+  constructor(
+    private readonly spaceAboutService: SpaceAboutService,
+    private spaceLookupService: SpaceLookupService
+  ) {}
 
   @AuthorizationAgentPrivilege(AuthorizationPrivilege.READ)
   @UseGuards(GraphqlGuard)
@@ -28,5 +35,23 @@ export class SpaceAboutResolverFields {
   ): Promise<IProfile> {
     const profile = await loader.load(space.id);
     return profile;
+  }
+
+  @AuthorizationAgentPrivilege(AuthorizationPrivilege.READ)
+  @ResolveField('metrics', () => [INVP], {
+    nullable: true,
+    description: 'Metrics about activity within this Space.',
+  })
+  async metrics(@Parent() space: ISpace) {
+    return await this.spaceAboutService.getMetrics(space);
+  }
+
+  @AuthorizationAgentPrivilege(AuthorizationPrivilege.READ)
+  @ResolveField('provider', () => IContributor, {
+    nullable: false,
+    description: 'The Space provider (host).',
+  })
+  async provider(@Parent() space: ISpace): Promise<IContributor> {
+    return await this.spaceLookupService.getProvider(space);
   }
 }
